[Subgraph] Refactor: Move Defined Behavior from `NetworkGraph` to `SubGraphBase`
- Moved the defined behavior from `NetworkGraph` to `SubGraphBase`.
- Updated public methods of `NetworkGraph` to pass their behaviors to `SubGraphBase`.
- Removed unused private methods in `NetworkGraph`.
