+++ nntrainer/graph/meson.build
@@ -1,6 +1,7 @@
 graph_sources = [
   'network_graph.cpp',
   'graph_core.cpp',
+  'subgraph_base.cpp',
   'connection.cpp'
 ]
 
@@ -9,6 +10,7 @@ graph_headers = [
   'graph_node.h',
   'network_graph.h',
   'graph_core.h',
+  'subgraph_base.h',
   'graph_node.h',
 ]
