+++ packaging/nntrainer.spec
@@ -603,6 +603,7 @@ cp -r result %{buildroot}%{_datadir}/nntrainer/unittest/
 %{_includedir}/nntrainer/model_common_properties.h
 %{_includedir}/nntrainer/network_graph.h
 %{_includedir}/nntrainer/graph_core.h
+%{_includedir}/nntrainer/subgraph_base.h
 %{_includedir}/nntrainer/manager.h
 %{_includedir}/nntrainer/basic_planner.h
 %{_includedir}/nntrainer/memory_planner.h
